feat: remove tenant related files for rhlc
